Simplify signature of remove_piece()
This is a non-functional simplification. Instead of passing the piece
type
for remove_piece, we can rely on the board. The only exception is
en-passant
which must be explicitly set because the destination square for the
capture
is not the same as the piece to remove.

Verified also in the Chess960 castling case by running a couple of
perft, see
the pull request discussion: #2460

STC
LLR: 2.94 (-2.94,2.94) [-3.00,1.00]
Total: 18624 W: 4147 L: 4070 D: 10407
Ptnml(0-2): 223, 1933, 4945, 1938, 260
http://tests.stockfishchess.org/tests/view/5dfeaa93e70446e17e451163

No functional change
master (#18)
@protonspring
@snicolet
protonspring authored