Simplify KPK classify
This is a non-functional simplification. If we use the "side to move" of
the entry
instead of the template, one of the classify methods goes away.
Furthermore, I've
resolved  the colors in some of the statements (we're already assuming
direction
using NORTH), and used stm (side to move) instead of "us," since this is
much clearer
to me.

This is not tested because it is non-functional, only applies building
the bitbase
and there are no changes to the binary (on my machine).

Closes #2485

No functional change
master (#18)
@protonspring
@snicolet
protonspring authored